requirements.txt
setuptools==68.2.2
wheel==0.41.2

flask==2.2.5
gunicorn==20.1.0
moviepy==1.0.3
imageio==2.31.1
numpy==1.24.4
tqdm==4.66.1
decorator==5.1.1
requests==2.31.0
google-api-python-client==2.98.0
google-auth==2.27.0
httplib2==0.22.0